Fix CI: stale iterdir preflight figure and a wasm spec loading the napi addon
The iterdir preflight now runs before the receiver copy is allocated, so the
reported usage in `iterdir_joins_are_preflighted` dropped by one path length.

The shared path checks move from `helpers.ts`, which imports the native
package at runtime, into `os_paths.ts`, which imports only types; each spec
passes its own backend's `MontyFileHandle`, so `npm run test:wasm` no longer
tries to load the napi addon.
