+++ crates/monty-js/__test__/helpers.ts
@@ -4,8 +4,7 @@
 
 import { afterAll as afterEachFile, beforeAll as beforeEachFile } from 'vitest'
 import { kind } from './env.js'
-import { Monty, MontyFileHandle, type CheckoutOptions, type FeedOptions } from '@pydantic/monty'
-import { t } from './assertions.js'
+import { Monty, type CheckoutOptions, type FeedOptions } from '@pydantic/monty'
 
 /** Checkout-level and feed-level options, flattened for convenience. */
 export interface RunOptions extends FeedOptions, CheckoutOptions {}
@@ -17,95 +16,6 @@ export interface PoolFixture {
   pool: () => Monty
 }
 
-/** Checks relative Python results while callbacks continue to receive absolute paths. */
-export async function checkRelativePathResults(run: PoolFixture['run']): Promise<void> {
-  const calls: unknown[] = []
-  const result = await run(
-    `from pathlib import Path
-([str(p) for p in Path('.').iterdir()],
- [str(p) for p in Path('sub/..').iterdir()],
- open('./file.txt').name,
- Path('./file.txt').open().name,
- str(open(b'./file.txt').name))`,
-    {
-      cwd: '/data',
-      os: (name, args) => {
-        calls.push([name, args])
-        if (name === 'Path.iterdir') return ['/data/file.txt']
-        if (name === 'open') return new MontyFileHandle(args[0] as string, 'r')
-        throw new Error(`unexpected OS call: ${name}`)
-      },
-    },
-  )
-  t.deepEqual(result, [['file.txt'], ['sub/../file.txt'], './file.txt', 'file.txt', "b'./file.txt'"])
-  t.deepEqual(calls, [
-    ['Path.iterdir', ['/data']],
-    ['Path.iterdir', ['/data']],
-    ['open', ['/data/file.txt', 'r']],
-    ['open', ['/data/file.txt', 'r']],
-    ['open', ['/data/file.txt', 'r']],
-  ])
-}
-
-/** Checks path rejection and error spelling through native and WASM OS callbacks. */
-export async function checkOsPathValidation(run: PoolFixture['run']): Promise<void> {
-  const calls: unknown[] = []
-  const result = await run(
-    `import os
-from pathlib import Path
-errors = []
-for operation in [
-    lambda: open(path),
-    lambda: Path(path).read_text(),
-    lambda: os.chdir(path),
-    lambda: os.rename(path, 'dst'),
-    lambda: os.rename('src', path),
-]:
-    try:
-        operation()
-    except ValueError as e:
-        errors.append(str(e))
-p = Path(path)
-(errors, p.exists(), p.is_file(), p.is_dir(), p.is_symlink(), os.getcwd())`,
-    {
-      cwd: '/data',
-      inputs: { path: 'bad\0/../x' },
-      os: (...args) => {
-        calls.push(args)
-        return true
-      },
-    },
-  )
-  t.deepEqual(result, [
-    [
-      'embedded null byte',
-      'embedded null byte',
-      'stat: embedded null character in path',
-      'rename: embedded null character in src',
-      'rename: embedded null character in dst',
-    ],
-    false,
-    false,
-    false,
-    false,
-    '/data',
-  ])
-  t.deepEqual(calls, [])
-  for (const [code, expected] of [
-    ['import os\nos.listdir()', "PermissionError: Permission denied: '/'"],
-    ["open('./x')", "PermissionError: Permission denied: '/x'"],
-    ["open('')", "PermissionError: Permission denied: ''"],
-    ["open('bad\\0/../x')", 'ValueError: embedded null byte'],
-  ] as const) {
-    const error = await t.throwsAsync(() => run(code, { cwd: '/' }))
-    t.is(error.message, expected)
-  }
-}
-
-/**
- * Registers before/after hooks creating and closing the spec file's shared
- * pool, and returns the `run` helper bound to it.
- */
 export function setupPool(): PoolFixture {
   let pool: Monty | null = null
   beforeEachFile(async () => {
